+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.31] - 2025-12-26
+
+- ✨ Add Extract Lines by Length (Equal, Less, Greater)
+- 🐛 Fix README and package.json description
+
 ## [0.0.30] - 2025-12-26
 
 - 🐛 Fix Half-width Katakana conversion